refactor(db): move user and DM store ownership
